refactor(services): migrate test imports to shared modules
Phase 2c: Update test imports and mock patches after bridge file deletion.

Changes:
- MIGRATE: 6 test files to use shared.* imports
- UPDATE: 10 mock patch targets from branch_arg to services
- UPDATE: architecture guard test to check shared.errors import

All tests pass (3094 passed).
