+++ tests/vibe3/commands/test_handoff_advanced_commands.py
@@ -59,7 +59,7 @@ class TestHandoffAdvancedCommands:
         assert result.exit_code == 0
 
     @patch("vibe3.commands.handoff_write.HandoffService")
-    @patch("vibe3.services.branch_arg.FlowService")
+    @patch("vibe3.services.FlowService")
     def test_handoff_append_command(self, mock_flow_service_class, mock_service_class):
         """Test handoff append command."""
         mock_flow_service = MagicMock()
@@ -93,7 +93,7 @@ class TestHandoffAdvancedCommands:
         )
 
     @patch("vibe3.commands.handoff_write.HandoffService")
-    @patch("vibe3.services.branch_arg.FlowService")
+    @patch("vibe3.services.FlowService")
     def test_handoff_plan_command(self, mock_flow_service_class, mock_service_class):
         """Test handoff plan command."""
         mock_flow_service = MagicMock()
@@ -124,7 +124,7 @@ class TestHandoffAdvancedCommands:
         )
 
     @patch("vibe3.commands.handoff_write.HandoffService")
-    @patch("vibe3.services.branch_arg.FlowService")
+    @patch("vibe3.services.FlowService")
     def test_handoff_report_command(self, mock_flow_service_class, mock_service_class):
         """Test handoff report command."""
         mock_flow_service = MagicMock()
@@ -155,7 +155,7 @@ class TestHandoffAdvancedCommands:
         )
 
     @patch("vibe3.commands.handoff_write.HandoffService")
-    @patch("vibe3.services.branch_arg.FlowService")
+    @patch("vibe3.services.FlowService")
     def test_handoff_audit_command(self, mock_flow_service_class, mock_service_class):
         """Test handoff audit command."""
         mock_flow_service = MagicMock()
@@ -255,9 +255,7 @@ class TestHandoffAdvancedCommands:
         ]
         mock_flow_service.store = mock_store
 
-        with patch(
-            "vibe3.services.branch_arg.FlowService", return_value=mock_flow_service
-        ):
+        with patch("vibe3.services.FlowService", return_value=mock_flow_service):
             result = runner.invoke(
                 app,
                 [
@@ -298,7 +296,7 @@ def test_handoff_next_sets_next_step_for_numeric_branch() -> None:
 
     with (
         patch("vibe3.commands.handoff_write.HandoffService", return_value=service),
-        patch("vibe3.services.branch_arg.FlowService", return_value=mock_flow_service),
+        patch("vibe3.services.FlowService", return_value=mock_flow_service),
     ):
         result = runner.invoke(
             app,
@@ -339,7 +337,7 @@ def test_handoff_next_rejects_nonexistent_flow() -> None:
 
     with (
         patch("vibe3.commands.handoff_write.HandoffService", return_value=service),
-        patch("vibe3.services.branch_arg.FlowService", return_value=mock_flow_service),
+        patch("vibe3.services.FlowService", return_value=mock_flow_service),
     ):
         result = runner.invoke(
             app,
