+++ tests/vibe3/commands/test_handoff_basic_commands.py
@@ -17,7 +17,7 @@ class TestHandoffBasicCommands:
 
     @pytest.mark.parametrize("force,expected_force", [(False, False), (True, True)])
     @patch("vibe3.commands.handoff_write.HandoffService")
-    @patch("vibe3.services.branch_arg.FlowService")
+    @patch("vibe3.services.FlowService")
     def test_handoff_init_command(
         self, mock_flow_service_class, mock_service_class, force, expected_force
     ):
