+++ tests/vibe3/services/test_shared_errors.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from unittest.mock import MagicMock, patch
 
 from vibe3.execution.contracts import ExecutionLaunchResult
-from vibe3.services.error_helpers import record_dispatch_failure_if_unexpected
+from vibe3.services.shared.errors import record_dispatch_failure_if_unexpected
 
 
 class TestRecordDispatchFailureIfUnexpected:
